requirements.txt
certifi==2020.4.5.1
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.3
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
olefile==0.46
opencv-python==4.2.0.34
Pillow==7.1.1
pyparsing==2.4.6
python-dateutil==2.8.1
six==1.14.0
tornado==6.0.4
wincertstore==0.2
